fixed success test Анализ производительности Текущие показатели:
Скорость: 79.86 док/сек (против 70 док/сек в v3 - прирост ~14%)

GPU загрузка: 87-88% - отлично!

Память GPU: 3.4-3.9 ГБ из 12-16 ГБ - есть запас

Очереди: q1=100, q2=1000, q3=0 - GPU успевают обрабатывать

Дисбаланс GPU:

GPU-0 (5060 Ti): 51,488 чанков, 29,634 док

GPU-1 (3060): 28,752 чанков, 17,501 док

Соотношение ~1.8:1 (должно быть ~1.5:1 при batch_size 32 vs 24)

+++ app/v4/gpu_worker.py
@@ -235,10 +235,10 @@ class GPUWorker(StoppableThread):
             padded_attention_mask.append(mask + [0] * pad_len)
         
         # Преобразуем в тензоры на GPU
-        input_ids = torch.tensor(padded_input_ids, device=self.device)
-        attention_mask = torch.tensor(padded_attention_mask, device=self.device)
+        input_ids = torch.tensor(padded_input_ids, device=self.device, dtype=torch.long)
+        attention_mask = torch.tensor(padded_attention_mask, device=self.device, dtype=torch.long)
         
-        # Инференс
+        # Инференс (модель сама приведёт эмбеддинги к своему типу)
         with torch.no_grad():
             outputs = self.model(input_ids=input_ids, attention_mask=attention_mask)
         
@@ -254,7 +254,7 @@ class GPUWorker(StoppableThread):
             chunk_len = len(chunk['input_ids'])
             chunk_preds = predictions[i, :chunk_len].cpu().tolist()
             
-            # Извлекаем сущности (используем проверенную функцию из прототипа)
+            # Извлекаем сущности
             entities = self._extract_entities_from_chunk(
                 chunk=chunk,
                 predictions=chunk_preds,
